chore: move botExecutor logic to framework
