+++ packages/create-botonic-app/dev-template/api/src/handlers/botExecutor.js
@@ -11,25 +11,5 @@ const dataProvider = dataProviderFactory(process.env.DATA_PROVIDER_URL)
 
 const dispatchers = dispatchersFactory(ENV, { sender })
 
-async function botExecutor({ userId, input, session, botState, websocketId }) {
-  const output = await bot.input({
-    input,
-    session,
-    botState,
-    dataProvider,
-  })
-  const events = [
-    ...output.messageEvents,
-    { action: 'update_bot_state', ...output.botState },
-    { action: 'update_session', ...output.session },
-  ]
-  // post events to sender sqs
-  await dispatchers.run('sender', {
-    userId,
-    events,
-    websocketId,
-  })
-}
-
 // eslint-disable-next-line no-undef
-export default botExecutorHandlerFactory(ENV, botExecutor)
+export default botExecutorHandlerFactory(ENV, bot, dataProvider, dispatchers)
