+++ packages/create-botonic-app/dev-template/api/src/handlers/sender.js
@@ -1,3 +1,7 @@
 import { senderHandlerFactory } from '@botonic/api/src/handlers/sender'
+import { dataProviderFactory } from '@botonic/core/lib/esm/data-provider'
+
+const dataProvider = dataProviderFactory(process.env.DATA_PROVIDER_URL)
+
 // eslint-disable-next-line no-undef
-export default senderHandlerFactory(ENV)
+export default senderHandlerFactory(ENV, dataProvider)
